+++ jdk/test/sun/tools/jcmd/TestJcmdSanity.java
@@ -35,15 +35,14 @@ import jdk.testlibrary.OutputAnalyzer;
 import jdk.testlibrary.ProcessTools;
 import jdk.testlibrary.Utils;
 
-/**
- * Unit test for jcmd utility. The test will send different diagnostic command
- * requests to the current java process.
- */
 /*
  * @test
  * @bug 7104647 7154822
+ * @summary Unit test for jcmd utility. The test will send different diagnostic
+ * command requests to the current java process.
+ *
  * @library /lib/testlibrary
- * @modules java.management
+ *
  * @build jdk.testlibrary.*
  * @run main/othervm -XX:+UsePerfData TestJcmdSanity
  */
@@ -71,7 +70,7 @@ public class TestJcmdSanity {
 
         output.shouldHaveExitValue(0);
         output.shouldNotContain("Exception");
-        output.shouldContain(Integer.toString(ProcessTools.getProcessId()) + ":");
+        output.shouldContain(Long.toString(ProcessTools.getProcessId()) + ":");
         matchJcmdCommands(output);
         output.shouldContain("For more information about a specific command use 'help <command>'.");
     }
